backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart==0.0.20
python-dotenv==1.0.1
uv==0.8.15
httpx==0.28.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
alembic==1.13.3
qdrant-client==1.12.0
typer==0.15.1
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
rich==13.9.4
sentence-transformers==3.0.1
openai==1.35.10
anthropic==0.34.1
google-generativeai==0.7.2